git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/trunk@1305838 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/core/src/java/org/apache/lucene/store/RAMOutputStream.java
@@ -177,7 +177,7 @@ public class RAMOutputStream extends IndexOutput {
 
   /** Returns byte usage of all buffers. */
   public long sizeInBytes() {
-    return file.numBuffers() * BUFFER_SIZE;
+    return (long) file.numBuffers() * (long) BUFFER_SIZE;
   }
   
   @Override
